Move files, add NodeClearInput, refacto
