+++ src/views/components/world/event/EventMenu.tsx
@@ -7,14 +7,14 @@ import { NewFolderButtonOnlyIcon, SecondaryButtonWithPlusIcon } from '@component
 import { NavigationDatabaseGroupStyle } from '@components/database/navigation/NavigationDatabaseGroup/NavigationDatabaseGroupStyle';
 import { useDialogsRef } from '@hooks/useDialogsRef';
 import { SeparatorGreyLine } from '@components/separators/SeparatorGreyLine';
-import { DbSymbol } from '../../../../models/entities/dbSymbol';
-import { useSetProjectText } from '../../../../utils/ReadingProjectText';
+import { DbSymbol } from '@modelEntities/dbSymbol';
+import { useSetProjectText } from '@utils/ReadingProjectText';
 import { EventTree } from './treeEvent/EventTree';
 import { EventEditorAndDeletionKeys, EventTreeEditorOverlay } from './editors/EventEditorOverlay';
-import { useEventTree } from '../../../../hooks/useEventTree';
-import { EVENT_FOLDER_NAME_TEXT_ID } from '../../../../models/entities/event/event-tree';
-import { addNewEventTreeFolder, findEventHasFolder } from '../../../../utils/events/EventUtils';
-import { useProjectEvents } from '../../../../hooks/useProjectData';
+import { useEventTree } from './hooks/useEventTree';
+import { EVENT_FOLDER_NAME_TEXT_ID } from '@modelEntities/event/event-tree';
+import { addNewEventTreeFolder, findEventHasFolder } from '@utils/events/EventTreeUtils';
+import { useProjectEvents } from '@hooks/useProjectData';
 
 const EventMenuContainer = styled(NavigationDatabaseStyle)`
   ${NavigationDatabaseGroupStyle} {
