+++ src/views/components/world/event/commands/CommandNodeProps.ts
@@ -1,10 +1,10 @@
 import type { StudioEventCommand, StudioEventCommandData } from '@modelEntities/event/command';
-import type { EventDialogsRef } from './editors/EventEditorOverlay';
+import type { CommandDialogsRef } from './editors/CommandEditorOverlay';
 
-export type EventNodeProps = {
+export type CommandNodeProps = {
   id: string;
   data: {
-    dialogsRef?: EventDialogsRef;
+    dialogsRef?: CommandDialogsRef;
     command: StudioEventCommandData<StudioEventCommand>;
     comments: string[];
   };
