+++ src/views/components/world/event/commands/InsertScriptCommand.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { EVENT_COMMAND_INSERT_SCRIPT_VALIDATOR, StudioEventCommandData, type StudioEventCommandInsertScript } from '@modelEntities/event/command';
-import { EventNodeProps } from './EventNodeProps';
+import { CommandNodeProps } from './CommandNodeProps';
 import { useCommandNode } from '../hooks/useCommandNode';
 import { useZodForm } from '@src/hooks/useZodForm';
 import { useNodeInputAttrsWithLabel } from '@src/hooks/useInputAttrs';
@@ -9,13 +9,12 @@ import React, { useEffect } from 'react';
 
 const INSERT_SCRIPT_EDITOR_SCHEMA = EVENT_COMMAND_INSERT_SCRIPT_VALIDATOR.pick({ script: true });
 
-export const InsertScriptCommand = ({ id, data: { dialogsRef, command, comments }, selected }: EventNodeProps) => {
+export const InsertScriptCommand = ({ id, data: { dialogsRef, command, comments }, selected }: CommandNodeProps) => {
   const { CommandNode, updateCommand } = useCommandNode<StudioEventCommandInsertScript>(id);
   const { type: commandType, ...commandData } = command as StudioEventCommandData<StudioEventCommandInsertScript>;
   const { canClose, getFormData, reload, defaults, formRef } = useZodForm(INSERT_SCRIPT_EDITOR_SCHEMA, commandData);
   const { MultiLineInput } = useNodeInputAttrsWithLabel(INSERT_SCRIPT_EDITOR_SCHEMA, defaults);
   const { t } = useTranslation();
-  const date = new Date().toLocaleString();
 
   const onBlur = () => {
     const result = canClose() && getFormData();
@@ -31,7 +30,6 @@ export const InsertScriptCommand = ({ id, data: { dialogsRef, command, comments 
 
   return (
     <CommandNode commandType={commandType} commentCount={comments.length} dialogsRef={dialogsRef} nodeId={id} selected={selected}>
-      <span>{date}</span>
       <InputFormContainer ref={formRef} onBlur={onBlur}>
         <MultiLineInput name="script" label={t(`event_command_script`)} className="nodrag" />
       </InputFormContainer>
