+++ src/views/components/world/event/commands/editors/DefaultEditor.tsx
@@ -1,15 +1,12 @@
 import { Editor } from '@components/editor';
 import { EditorHandlingClose, useEditorHandlingClose } from '@components/editor/useHandleCloseEditor';
-import { useGlobalState } from '@src/GlobalStateProvider';
 import React, { forwardRef } from 'react';
 
 export const DefaultEditor = forwardRef<EditorHandlingClose>((_, ref) => {
-  const [, setState] = useGlobalState();
   const canClose = () => true;
 
   const onClose = () => {
-    //console.log('bye bye');
-    setState((s) => ({ ...s, textVersion: s.textVersion + 1 }));
+    console.log('bye bye');
   };
   useEditorHandlingClose(ref, onClose, canClose);
