+++ src/views/components/world/event/common/CustomEdge.tsx
