+++ src/views/components/world/event/common/EventIcon.tsx
