+++ src/views/components/world/event/common/ShadowNode.tsx
