+++ src/views/components/world/event/editors/EventFolderDeletion.tsx
@@ -3,8 +3,8 @@ import { EditorHandlingClose, useEditorHandlingClose } from '@components/editor/
 import { DbSymbol } from '@modelEntities/dbSymbol';
 import { useGetEntityNameTextUsingTextId } from '@utils/ReadingProjectText';
 import { useProjectEvents } from '@hooks/useProjectData';
-import { useEventTree } from '@hooks/useEventTree';
-import { getEventTreeChildrenDbSymbols, removeEventTreeItem } from '@utils/events/EventUtils';
+import { useEventTree } from '@components/world/event/hooks/useEventTree';
+import { getEventTreeChildrenDbSymbols, removeEventTreeItem } from '@utils/events/EventTreeUtils';
 import React, { forwardRef, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
